+++ resources/views/cards/analytics-campaigns.blade.php
@@ -15,11 +15,15 @@
                     <div class="tui-analytics-col">
                         <h4 class="tui-analytics-col-title">{{ $section['title'] }}</h4>
                         <table class="tui-table">
+                            <thead>
+                                <tr><th></th><th class="is-num">Views</th><th class="is-num">Visitors</th></tr>
+                            </thead>
                             <tbody>
                                 @foreach ($section['rows'] as $row)
                                     <tr>
                                         <td class="is-primary">{{ $row['key'] }}</td>
                                         <td class="is-num">{{ Format::count($row['views']) }}</td>
+                                        <td class="is-num tui-tone-dim">{{ Format::count($row['visitors']) }}</td>
                                     </tr>
                                 @endforeach
                             </tbody>